[rename] #85 - API 파일 이름 수정
